add WorkHeader component
